### 2020-05-28 #2270 @whedon
[PRE REVIEW]: GPP, the Generic Preprocessor
**Submitting author:** @Logological (<a href="http://orcid.org/0000-0002-0749-1100">Tristan Miller</a>)
**Repository:** <a href="https://github.com/logological/gpp.git" target ="_blank">https://github.com/logological/gpp.git</a>
**Version:** 2.26
