+++ paper.bib
@@ -287,11 +287,12 @@
 @article{triance1980structured,
   author       = {Triance, J. M.},
   title	       = {Structured Programming in {COBOL}—The Current Options},
-  journal      = {The Computer Journal},
   year	       = 1980,
   volume       = 23,
   number       = 3,
   pages	       = {194--200},
+  doi	       = {10.1093/comjnl/23.3.194},
+  journal      = {The Computer Journal},
 }
 
 @article{walden2014macro,
