### 2020-05-28 #2270 @gkthiruvathukal
Hi, @logological, I'll be serving as the editor for this submission. About to assign myself. I will be asking you to suggest potential reviewers from our list (see the top of this the thread).
