### 2020-06-15 45e028e
Clarify that only .tar.bz2 are portable source packages
