### 2020-06-18 #2270 @gkthiruvathukal
@rljacobson and smattr, would you be willing to serve as reviewers for this JOSS submission?
