### 2020-06-18 #2270 @gkthiruvathukal
@smattr, oops, I forgot the mention symbol...  would you be willing to serve as reviewers for this JOSS submission?
