### 2020-06-19 #2270 @Smattr
Yes, certainly. I'm not qualified to review the Pascal, but I can review the C. I've had a lot of experience with CPP and M4, so I should also be able to review the usability/functionality too.
