### 2020-06-19 #2270 @logological
@Smattr, don't be misled by whedon's automated analysis, which is full of false positives. The software described in the paper is written entirely in C.  There's no Pascal, no C++, and no M4 (except insofar as that's what the Autotools configuration files are processed with, though you won't need to touch those if you build from the portable source packages).
