### 2020-06-26 #2270 @davidbeckingsale
> @davidbeckingsale, would you be willing to serve as reviewers for this JOSS submission?

I don't have the bandwidth for this right now, so will decline. Thanks for considering me.
