### 2020-06-27 #2270 @gkthiruvathukal
@Smattr That's correct. You are certainly welcome to start looking things over until the review thread is formally opened. Going to try to get the second reviewer assigned. Thanks!
