### 2020-06-27 #2270 @drj11
in principle yes I can review this JOSS submission. It would be my first review for JOSS. Are there guidelines for reviewers?
