### 2020-06-27 #2270 @gkthiruvathukal
@drj11 Thanks for responding. As a matter of fact, we do have guidelines at https://joss.readthedocs.io/en/latest/reviewer_guidelines.html?

I hope I can interpret your response as a "yes". I'll be happy to help with any questions during the process.
