### 2020-07-12 #2400 @Smattr
Certainly. I made progress on a few of the initial checklist items, as indicated above. However, unfortunately I didn't have much time over the last couple of weeks to do more than that. I should have time this coming week to get deeper into the review.
