### 2020-07-12 #2400 @Smattr
A couple of questions:
1. GPP appears to have no test suite, which is one of the items on the review checklist. However, in my experience it is extremely challenging to define a relevant suite of test cases for a tool as generic as this. In practice, taking bug reports from your users based on the bizarre and creative things they attempt with your tool is more fruitful and representative of real world scenarios. Is this item a strict requirement? @logological, maybe you can comment on the choice to omit a test suite too?
2. Does Whedon check the reference syntax? The PDF output looks fine to me, but the link for that item in the checklist goes to some advice about R Markdown suggesting to me that I'm supposed to review the source rather than the PDF.
