### 2020-07-13 #2400 @logological
@smattr: According to the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html), a fully automated test suite is not required.  It is acceptable to provide instead "documented manual steps that can be followed to objectively check the expected functionality of the software (e.g., a sample input file to assert behaviour)".  I believe that we've met this requirement by including in the distribution a rather large example input file and its output—namely, the GPP documentation itself.  As we write in the README and on the GPP website,

> Note that the documentation itself is generated by GPP using an input file (`gpp.pp`) with macros that can automatically produce HTML, LaTeX, or troff (man page) output. Studying this file should prove instructive, and manually running it through GPP can serve as a basic test of the program's functionality.

Besides this, the documentation includes a large number of example code snippets, most of which are accompanied by the expected output or a prose description thereof.

Providing an automated test suite using Automake is something that I'd eventually like to do, though this will take some effort and, given what I've written above, I didn't think it necessary to do so for the JOSS submission.
