### 2020-07-13 #2400 @drj11
@gkthiruvathukal The checkbox for LICENSE requires specifically a `LICENSE` file. This repository has a `COPYING` file, which clearly fulfills the same purpose (it is an older, GPL-era, convention). Should I file an issue to copy the text to `LICENSE`? This reviewer would probably prefer that the guidelines accept either.
