### 2020-07-13 logological/gpp#38 @drj11
There is no INSTALL file
Per my review at https://github.com/openjournals/joss-reviews/issues/2400

For installation your README says "see the included INSTALL file." but there is no `INSTALL` file.
