### 2020-07-13 logological/gpp#38 @logological
The INSTALL file is automatically generated by the GNU Autotools when they produce the portable source packages.  As a generated file, it is not checked into the source repository.  If you download the portable source packages, you can verify that the file exists there.
