### 2020-07-13 logological/gpp#38 @drj11
Yes, that is what I'm doing now. In which case that's fine for the purposes of review, but I would encourage you to mention this fact in the README. The README does mention that you can clone the github repo, but perhaps it should not. Or it should say that cloning is only recommended for developers or people willing to get familiar with autotools.
