### 2020-07-13 logological/gpp#38 @logological
OK, I will make the README a bit more explicit in this respect.
