### 2020-07-13 715642e
Add warning about source repository. Fixes Issue #38.
