### 2020-07-13 logological/gpp#39 @drj11
Examples in HTML documentation suffer from an extra indent
The examples at https://files.nothingisreal.com/software/gpp/gpp.html are indented by two spaces in their PRE blocks.

This means that they cannot be cut and pasted from the HTML page to `gpp`. Or rather, they can, and it works, but I suspect that the output has extra spaces that are not intended:

Evidence:

```
root@f14987a42448:/gpp-2.26# cat ex0
  #define FOO This is
  #define BAR a message.
  #define concat #1 #2
  concat(FOO,BAR)
  #ifeq (concat(foo,bar)) (foo bar)
  This is output.
  #else
  This is not output.
  #endif
root@f14987a42448:/gpp-2.26# gpp ex0
        This is a message.
    This is output.
  root@f14987a42448:/gpp-2.26#
```

This won't stop my review at https://github.com/openjournals/joss-reviews/issues/2400 from passing, but is a small quality issue.
