### 2020-07-13 logological/gpp#40 @drj11
list of dependencies could be more explicit
Per my review on https://github.com/openjournals/joss-reviews/issues/2400

The checklist item asks "Is there a clearly-stated list of dependencies?"

I think there should be a more explicit list of dependencies. I don't think this needs to be long.

Where the README currently says: "(In brief, ./configure && make && make install should work in most cases.)" you could add "if you have `make` and a C compiler installed," after "in brief".

It's worth noting that neither OS X nor Ubuntu come with either make or a C compiler installed by default. You could add something like, "on Ubuntu `apt install build-essentials` should be sufficient", but I regard that as entirely optional, and it's up to you. I'm only suggesting Ubuntu because that's what I use, you probably have a better idea of what your users likely use.
