### 2020-07-13 logological/gpp#41 @drj11
Write out the test in the README
Per my review on https://github.com/openjournals/joss-reviews/issues/2400

I agree with your comment on https://github.com/openjournals/joss-reviews/issues/2400#issuecomment-657416699 : `gpp`'s documentation is a sufficient test of `gpp`. But It would be best for my conscience in checking the check box if this were more explicit.

I recommend:

- Put "GPP's own documentation uses GPP and generates HTML, LaTeX, and man pages (troff) from one `gpp` source file" in the first paragraph at the top of the README. It's a good feature of gpp and a useful demonstration of its capabilities.
- After the README's Documentation section add a Testing section: "`gpp -H -Dhtml doc/gpp.pp -o gpp.html` builds GPP's HTML document and serves as a basic test; for more see the `doc/gpp.pp` file".
