### 2020-07-13 #2400 @drj11
I am giving a generous amount of wiggle room for "non-specialist audience" because I don't believe that there really are any non-specialist audiences for macro pre-processors.
