### 2020-07-13 #2400 @gkthiruvathukal
@drj11 Thanks for getting started!  That's why I provide the gentle nudge.  Regarding the LICENSE file, my interpretation is that it is required for the file to be named LICENSE. I tend to agree that there are traditions (e.g. GPL-era) but, for a journal, we want to see some degree of consistency among projects. I'm guessing a simple `git mv` will do the trick.
