### 2020-07-13 #2400 @gkthiruvathukal
@Smattr and @logological, regarding test suites, I agree that a full test suite can be painful. However, it does seem like script that checks a number of inputs against expected outputs would be more than appropriate. Based on a read of @logological's follow up, it sounds like this is indeed the approach he has taken.

Having done a lot of work on language front ends in the past and accompanying tools (levers, parsers, ASTs), I think this approach is acceptable as long as we have a robust set of cases (inputs vs. expected outputs).
