### 2020-07-13 c9e5bc2
Documentation updates. Fixes #39, fixes #42, fixes #43, fixes #44.
