### 2020-07-13 #2400 @logological
> Regarding the LICENSE file, my interpretation is that it is required for the file to be named LICENSE. I tend to agree that there are traditions (e.g. GPL-era) but, for a journal, we want to see some degree of consistency among projects. I'm guessing a simple `git mv` will do the trick.

Yes, consistency is best.  But in this case it's JOSS which is the outlier: for 30-odd years it's been a de facto standard in the C-coding free software community (and a de jure standard for GNU packages) that licensing information goes in a file called `COPYING` (or `COPYING.LESSER` for LGPL-licensed projects).  The GNU Autotools, which is probably still the most common packaging system for C software on *nix, hard-code this filename convention.  I can't simply rename `COPYING` to `LICENSE` without changing my build scripts; this will moreover confuse and/or irritate many users who are expecting the licence to be in the standard place, and will prevent me from submitting GPP as an official GNU package (which is something I have been mulling over for a while).

I'd ask that JOSS's rule about the licence file be interpreted liberally in this case, or else globally changed to reflect the reality that different programming languages and communities have different standards and traditions concerning the naming of the licence file.
