### 2020-07-13 #2400 @drj11
I agree with @logological on the COPYING/LICENSE here. While GitHub is encouraging people to add LICENSE files, we should accept long-standard community practice that share the same goals. I note that the GitHub sidebar "knows" that the project is LGPL-3.0 licensed and links to the file called `COPYING`. So, even though GitHub documentation tell you to create a file called LICENSE, it still accepts and knows about it if you put it in a file called `COPYING`.

Would a symbolic link be acceptable?
