### 2020-07-13 #2400 @gkthiruvathukal
@logological and @drj11,  I tend to agree this request merits consideration.

@openjournals/joss-eics  Please see the above regarding the LICENSE file issue. Our review guidelines don't expressly prohibit placing he license in COPYING, which is really part of the GNU/GPL tradition. What is the EICs' position on this?
