### 2020-07-13 #2400 @drj11
It's not really JOSS that's the outlier (or at least, JOSS is not alone in this outlier), it's really "the non-GNU internet". But @logological is right about 30 years of GNU-and-others tradition.
