### 2020-07-13 #2400 @logological
> Yes, consistency is best. But in this case it's JOSS which is the outlier: for 30-odd years it's been a de facto standard in the C-coding free software community (and a de jure standard for GNU packages) that licensing information goes in a file called `COPYING` (or `COPYING.LESSER` for LGPL-licensed projects).

Oh, and just to give some quick third-party evidence of this: I surveyed the `/usr/share/doc` directory tree for the official packages installed on my openSUSE system; `COPYING*` files outnumber `LICENSE*` files by about two to one.  Packages using `COPYING` are overwhelmingly C and C++ programs and libraries, while those using `LICENSE` tend to be written in Ruby, Rust, Perl, etc., or else do not contain source code at all (i.e., font or dictionary packages).
