### 2020-07-13 #2400 @drj11
The section on licensing in [the review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) says:

> Software license
>
> There should be an OSI approved license included in the repository. Common licenses such as those listed on choosealicense.com are preferred. Note there should be an actual license file present in the repository not just a reference to the license.
>
>     Acceptable: A plain-text LICENSE file with the contents of an OSI approved license
>     Not acceptable: A phrase such as ‘MIT license’ in a README file

And this repository clearly meets the criteria even if it does not use a file called LICENSE. I think the checkpoint text should be changed to "Does the repository contain a plain-text license file with the contents of an OSI approved software license? Ideally in the file called `LICENSE` but other community-recognised names, such as `COPYING`, are also acceptable.".
