### 2020-07-13 10500b7
Update installation instructions. Fixes #40.
