### 2020-07-13 8aabf89
Add instructions on testing GPP. (Issue #41)
