+++ README.md
@@ -61,10 +61,12 @@ man page and an HTML file.  The HTML version is also available online
 on the [GPP home page](https://logological.org/gpp).
 
 Note that the documentation itself is generated by GPP using an input
-file (`gpp.pp`) with macros that can automatically produce HTML,
-LaTeX, or troff (man page) output.  Studying this file should prove
-instructive, and manually running it through GPP can serve as a basic
-test of the program's functionality.
+file ([`gpp.pp`](doc/gpp.pp)) with macros that can automatically
+produce HTML, LaTeX, or troff (man page) output.  Studying this file
+should prove instructive, and manually running it through GPP can
+serve as a basic test of the program's functionality.  Once you have
+installed GPP, you can invoke it as follows to get instructions on how
+to manually generate the documentation: `gpp -H /path/to/gpp.pp`
 
 
 Contributing
