### 2020-07-13 logological/gpp#41 @logological
Thanks for the suggestion.  I'd prefer not to mention this in the lede paragraph as I don't consider it central to the description of what GPP is.  However, I've expanded the part of the README where I mention generating the documentation as a test of GPP's capabilities; as you suggested I explicitly gave the command to run and explicitly referenced the `doc/gpp.pp` file (with a hyperlink).  Note that I didn't specifically instruct the user to generate the HTML version of the documentation, since running `gpp -H gpp.pp` without any further arguments will output a help message with the exact command lines to use to generate the troff, HTML, and LaTeX documentation:

```
 To get the man page, run:
    gpp -H -Dman gpp.pp -o gpp.1

 To get the HTML page, run:
    gpp -H -Dhtml gpp.pp -o gpp.html

 To get a LaTeX version, run:
    gpp -H -Dlatex gpp.pp -o gpp.tex
```

I've also opened a new issue (Issue #45) for a proper test suite, though as I mentioned in the review issue, I think this can be left as future work.  (Hopefully I can get around to this some time in the next few weeks or months.)

I hope this is sufficient for you—if not, feel free to reopen this issue.
