### 2020-07-13 #2400 @gkthiruvathukal
@drj11 and @logological, it looks like our discussion has already resulted in a patch! So please proceed with the rest of the review. COPYING is a valid license file now!
