### 2020-07-13 #2400 @drj11
I have the same References question as @Smattr . Since ·whedon helpfully compiles the PDF, should we be looking at the references as they appear in the PDF and not the `.bib` file?
