### 2020-07-13 #2400 @gkthiruvathukal
@drj11 Yes, please check both the .bib and the PDF. If there is a problem with the entry or its rendering, it needs to be reported as an issue.
