### 2020-07-21 #2400 @drj11
I have checked paper.bib using https://biblatex-linter.herokuapp.com/validate which notes that 4 entries are missing an institution. In my opinion this is of no consequence.
