### 2020-07-21 #2400 @logological
> I have checked paper.bib using https://biblatex-linter.herokuapp.com/validate which notes that 4 entries are missing an institution. In my opinion this is of no consequence.

This is because that validator expects its input to use the biblatex dialect, whereas my bibliography entries are written in the older BibTeX dialect.  Whatever program JOSS is using to compile the paper is correctly recognizing the entries as BibTeX and is properly including the institutional information (which is in the `school` field).
