### 2020-07-21 #2400 @drj11
@logological ah, that is reassuring. I was wondering how I, having last touched a bitex file in 1992, was expected to know the difference between the two. :)
