### 2020-07-21 #2400 @logological
I was wondering myself which dialect to use for the bibliography, since I don't think this is covered in the submission instructions.  I decided it was safer to go with BibTeX since the modern biblatex bibliographic processor can handle both biblatex and BibTeX dialects, whereas the older BibTeX processor doesn't understand biblatex.  Perhaps this could be clarified in the JOSS style guide so that authors don't have to guess like I did.
