### 2020-07-23 d4c4ec3
Print --help and --version messages to stdout. Fixes #47.
