+++ src/gpp.c
@@ -336,43 +336,42 @@ void PopSpecs(void) {
 }
 
 void display_version(void) {
-    fprintf(stderr, PACKAGE_STRING "\n");
-    fprintf(stderr, "Copyright (C) 1996-2001 Denis Auroux\n");
-    fprintf(stderr, "Copyright (C) 2003-2020 Tristan Miller\n");
-    fprintf(stderr,
-            "This is free software; see the source for copying conditions.  There is NO\n"
-            "warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.\n");
+    printf(PACKAGE_STRING "\n");
+    printf("Copyright (C) 1996-2001 Denis Auroux\n");
+    printf("Copyright (C) 2003-2020 Tristan Miller\n");
+    printf("This is free software; see the source for copying conditions.  There is NO\n"
+           "warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.\n");
 }
 
 void usage(void) {
-    fprintf(stderr,"Usage : gpp [-{o|O} outfile] [-I/include/path] [-Dname=val ...] [-z] [-x] [-m]\n");
-    fprintf(stderr,"            [-n] [-C | -T | -H | -X | -P | -U ... [-M ...]] [+c<n> str1 str2]\n");
-    fprintf(stderr,"            [+s<n> str1 str2 c] [long options] [infile]\n\n");
-    fprintf(stderr,"      default:    #define x y           macro(arg,...)\n");
-    fprintf(stderr," -C : maximum cpp compatibility (includes -n, +c, +s, ...)\n");
-    fprintf(stderr," -T : TeX-like    \\define{x}{y}         \\macro{arg}{...}\n");
-    fprintf(stderr," -H : HTML-like   <#define x|y>         <#macro arg|...>\n");
-    fprintf(stderr," -X : XHTML-like  <#define x|y/>        <#macro arg|.../>\n");
-    fprintf(stderr," -P : prolog compatible cpp-like mode\n");
-    fprintf(stderr," -U : user-defined syntax (specified in 9 following args; see manual)\n");
-    fprintf(stderr," -M : user-defined syntax for meta-macros (specified in 7 following args)\n\n");
-    fprintf(stderr," -o : output to outfile\n");
-    fprintf(stderr," -O : output to outfile and stdout\n");
-    fprintf(stderr," -z : line terminator is CR-LF (MS-DOS style)\n");
-    fprintf(stderr," -x : enable #exec built-in macro\n");
-    fprintf(stderr," -m : enable automatic mode switching upon including .h/.c files\n");
-    fprintf(stderr," -n : send LF characters serving as macro terminators to output\n");
-    fprintf(stderr," +c : use next 2 args as comment start and comment end sequences\n");
-    fprintf(stderr," +s : use next 3 args as string start, end and quote character\n\n");
-    fprintf(stderr," Long options:\n");
-    fprintf(stderr," --include file : process file before infile\n");
-    fprintf(stderr," --nostdinc : don't search standard directories for files to include\n");
-    fprintf(stderr," --nocurinc : don't search the current directory for files to include\n");
-    fprintf(stderr," --curdirinclast : search the current directory last\n");
-    fprintf(stderr," --warninglevel n : set warning level\n");
-    fprintf(stderr," --includemarker formatstring : keep track of #include directives in output\n\n");
-    fprintf(stderr," --version : display version information and exit\n");
-    fprintf(stderr," -h, --help : display this message and exit\n\n");
[…]
