### 2020-07-25 #2400 @logological
@gkthiruvathukal Should I make a new tagged release of GPP incorporating the documentation changes resulting from the review?  Or are there any further changes you think are necessary before you think this submission can be accepted for publication?
