### 2020-07-26 #2400 @gkthiruvathukal
@logological Yes, it looks like we are ready for the next steps. Please do the following:

- [ ] Make a tagged release of your software, and list the version tag of the archived version here. [That answers your question above!]
