### 2020-07-27 870d1a4
Minor copy-editing of paper
