### 2020-07-27 #2400 @logological
OK, the new release tag in GitHub is as follows: 2.27
The Zendo DOI is as follows: 10.5281/zenodo.3961322
