### 2020-07-27 #2400 @gkthiruvathukal
It appears all of the items on my checklist are addressed, @logological.

@openjournals/jose-eics I'm ready to proceed with accepting this submission.
