### 2020-07-27 #2400 @gkthiruvathukal
@openjournals/joss-eics My bad, this is for JOSS, not JOSE. :)
