### 2020-07-28 #2400 @kthyng
Still waiting for the doi to resolve to close the issue, but in the meantime...

Congrats to  @logological on your new publication!! Thanks to @gkthiruvathukal for editing and to reviewers @Smattr, @drj11. This process couldn't happen without your time and expertise.
