### 2020-07-28 #2400 @kthyng
@openjournals/dev the doi still isn't resolving for me — can you take a look?
