### 2020-07-28 #2400 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
