### 2020-07-29 #2400 @logological
Thank you to the reviewers @drj11 and @Smattr for your many insightful comments and questions, and to @gkthiruvathukal for shepherding this first submission of mine to a speedy acceptance!
